requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.13.1
Flask==3.1.1
flask-cors==6.0.1
fsspec==2024.6.1
gdown==5.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
pillow==11.0.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.4
setuptools==70.2.0
soupsieve==2.7
sympy==1.13.3
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.4.0
Werkzeug==3.1.3
